# apps/api/src/rhapto/api/routers/tailor.py
from __future__ import annotations

import asyncio
import json
import uuid
from collections.abc import AsyncIterator
from typing import Annotated

from fastapi import APIRouter, Depends, HTTPException
from sqlalchemy.ext.asyncio import AsyncSession
from sse_starlette.sse import EventSourceResponse

from rhapto.api.deps import (
    AppState,
    current_user,
    get_enqueuer,
    get_event_bus,
    get_session,
    get_settings_dep,
    get_state,
    get_storage,
)
from rhapto.api.errors import not_found
from rhapto.api.schemas import TailorBody, TaskOut
from rhapto.config import Settings
from rhapto.db.models import Task
from rhapto.db.repositories import documents as documents_repo
from rhapto.db.repositories import jobs as job_repo
from rhapto.db.repositories import packages as package_repo
from rhapto.db.repositories import tasks as task_repo
from rhapto.db.repositories.profile import get_track, library_state
from rhapto.services.enqueue import Enqueuer
from rhapto.services.eventbus import EventBus, task_channel
from rhapto.services.llm import resolve_llm_config
from rhapto.services.storage import PackageStorage
from rhapto.services.trial import check_trial_allowance

router = APIRouter()
FINISHED = {"succeeded", "failed"}

UserDep = Annotated[uuid.UUID, Depends(current_user)]
SessionDep = Annotated[AsyncSession, Depends(get_session)]
StorageDep = Annotated[PackageStorage, Depends(get_storage)]
EnqueuerDep = Annotated[Enqueuer, Depends(get_enqueuer)]
SettingsDep = Annotated[Settings, Depends(get_settings_dep)]
EventBusDep = Annotated[EventBus, Depends(get_event_bus)]
StateDep = Annotated[AppState, Depends(get_state)]


def task_to_out(row: Task) -> TaskOut:
    return TaskOut(
        id=row.id,
        type=row.type,
        status=row.status,
        progress=dict(row.progress_json),
        error=row.error,
        result_ref=row.result_ref,
        created_at=row.created_at,
        finished_at=row.finished_at,
    )


@router.post("/jobs/{job_id}/tailor", response_model=TaskOut, status_code=202)
async def tailor_job_endpoint(
    job_id: uuid.UUID,
    body: TailorBody,
    user_id: UserDep,
    session: SessionDep,
    enqueuer: EnqueuerDep,
    settings: SettingsDep,
    storage: StorageDep,
) -> TaskOut:
    if await job_repo.get_job(session, user_id, job_id) is None:
        raise not_found("job", job_id)
    if body.track_id is not None and await get_track(session, user_id, body.track_id) is None:
        raise HTTPException(status_code=422, detail=f"unknown track {body.track_id!r}")
    if body.parent_package_id is not None:
        parent = await package_repo.get_package(session, user_id, body.parent_package_id)
        if parent is None or parent.job_id != job_id:
            raise not_found("package", body.parent_package_id)
    has_document = await documents_repo.get_document(session, user_id) is not None and (
        await asyncio.to_thread(storage.has_document, user_id)
    )
    if body.mode == "tune" and not has_document:
        raise HTTPException(
            status_code=422,
            detail="upload a resume document before tailoring in tune mode",
        )
    # Resolve the mode HERE, before the key check: the default is tune when a document is stored, so
    # checking only `body.mode == "blocks"` would miss the common case. Blocks mode needs a block
    # library and a track; saying so now beats a 202 whose task can only fail in the worker.
    mode = body.mode or ("tune" if has_document else "blocks")
    if mode == "blocks":
        library = await library_state(session, user_id)
        if not (library.blocks and library.tracks):
            raise HTTPException(
                status_code=422,
                detail=(
                    "Rhapto needs your experience before it can build a resume from your library. "
                    "Upload a resume document to tune it instead, or add your blocks and a track "
                    "in Profile."
                ),
            )
    # Fail here rather than leaving a task row that can only fail in the worker: without a provider
    # key there is nothing to run. LLMNotConfiguredError becomes the 409 the web app turns into
    # "set up your LLM in Settings" (see api.errors). Last of the checks, so a request that is also
    # malformed hears about that first.
    await resolve_llm_config(session, settings, user_id)
    # And then: is this account still allowed to spend the DEPLOYMENT's key? Read-only -- nothing is
    # consumed here and the worker's claim is the enforcement. Ordering is deliberate: a user with no
    # key anywhere keeps hearing the 409 above, so the trial refusal is unreachable on a deployment
    # that has no key of its own to protect. No task row, nothing enqueued: the person gets an
    # immediate, explainable answer instead of a task that fails a minute later.
    await check_trial_allowance(session, settings, user_id)
    # The worker loads the document itself (the bytes never travel through the task row), so
    # all the request carries is the resolved mode.
    request = {
        "job_id": str(job_id),
        "track_id": body.track_id,
        "feedback": body.feedback,
        "parent_package_id": str(body.parent_package_id) if body.parent_package_id else None,
        "mode": mode,
    }
    task = await task_repo.create_task(session, user_id, "tailor_job", {"request": request})
    await session.commit()  # the worker (inline or arq) must see the row
    task_id = task.id
    await enqueuer.enqueue("tailor_job", task_id=str(task_id))
    session.expire_all()
    refreshed = await task_repo.get_task(session, user_id, task_id)
    if refreshed is None:  # committed above, so this cannot happen without a concurrent delete
        raise RuntimeError(f"task {task_id} vanished after commit")
    return task_to_out(refreshed)


@router.get("/tasks/{task_id}", response_model=TaskOut)
async def get_task(task_id: uuid.UUID, user_id: UserDep, session: SessionDep) -> TaskOut:
    task = await task_repo.get_task(session, user_id, task_id)
    if task is None:
        raise not_found("task", task_id)
    return task_to_out(task)


@router.get("/tasks/{task_id}/events")
async def task_events(
    task_id: uuid.UUID,
    user_id: UserDep,
    bus: EventBusDep,
    state: StateDep,
) -> EventSourceResponse:
    # No request-scoped session here: a `SessionDep` would stay open (and hold its pooled
    # connection in an idle read transaction) for the whole life of the stream.
    async with state.session_factory() as check:
        if await task_repo.get_task(check, user_id, task_id) is None:
            raise not_found("task", task_id)

    async def stream() -> AsyncIterator[dict[str, str]]:
        async with bus.subscription(task_channel(str(task_id))) as events:
            # Read the state *after* subscribing, on its own short-lived session, so an event
            # published between the existence check and the subscribe cannot be missed.
            async with state.session_factory() as fresh:
                current = await task_repo.get_task(fresh, user_id, task_id)
            if current is None:  # deleted while we were subscribing
                return
            yield {"event": "state", "data": task_to_out(current).model_dump_json()}
            if current.status in FINISHED:
                return
            async for event in events:
                yield {"event": str(event.get("event", "message")), "data": json.dumps(event)}
                if event.get("event") in {"done", "error"}:
                    return

    # sse-starlette defaults to "\r\n" line endings; use "\n" so plain-"\n\n"
    # SSE block parsing (ours and typical browser EventSource clients) works.
    return EventSourceResponse(stream(), sep="\n")